Debloated unneccessary code
