--- defaults/ifetchrc
@@ -1,14 +1,0 @@
-# Logo color
-loc=G
-
-# Fields color
-fc=W
-
-# Separator
-s=" >> "
-
-# Separator color
-sc=G
-
-# Values color
-vc=W
