--- defaults/logos/eth
@@ -1,8 +1,0 @@
-+---------------+
-|   +-------+   |
-| +--       --+ |
-| |           | |
-| | | | | | | | |
-| | | | | | | | |
-| +-----------+ |
-+---------------+
