--- defaults/logos/fallback
@@ -1,8 +1,0 @@
-            +---+
-            |   |
-        +---+   |
-        |   |   |
-    +---+   |   |
-    |   |   |   |
-+---+   |   |   |
-|___|___|___|___|
