--- defaults/logos/wifi
@@ -1,8 +1,0 @@
-   ___________   
-  /           \  
- /  _________  \ 
-/  /         \  \
-  /  _______  \  
-    /       \    
-       ...       
-       ...
